+++ src/execution/mod.rs
@@ -4,7 +4,6 @@ pub mod data_context;
 pub mod oauth2_cache;
 pub mod oauth2_client_tokens;
 pub mod oauth2_pkce;
-pub mod results;
 pub mod tally;
 pub mod test_runner;
 
@@ -14,6 +13,5 @@ pub use data_context::*;
 pub use oauth2_cache::*;
 pub use oauth2_client_tokens::*;
 pub use oauth2_pkce::*;
-pub use results::*;
 pub use tally::*;
 pub use test_runner::*;